+++ pkg/render/helm.go
@@ -2,17 +2,26 @@ package render
 
 import (
 	"io"
+	"path/filepath"
+	"strings"
+
+	"helm.sh/helm/v3/pkg/chart"
 
 	fleet "github.com/rancher/fleet/pkg/apis/fleet.cattle.io/v1alpha1"
 	"github.com/rancher/fleet/pkg/bundlereader"
-	"github.com/rancher/fleet/pkg/helm"
+	"github.com/rancher/fleet/pkg/fleetyaml"
 	"github.com/rancher/fleet/pkg/manifest"
 	"github.com/rancher/fleet/pkg/patch"
+	"github.com/rancher/fleet/pkg/rawyaml"
+
+	"github.com/rancher/wrangler/pkg/kv"
+
+	"sigs.k8s.io/yaml"
 )
 
-// ToChart applies overlays to "manifest"-style gitrepos and transforms the
+// HelmChart applies overlays to "manifest"-style gitrepos and transforms the
 // manifest into a helm chart tgz
-func ToChart(name string, m *manifest.Manifest, options fleet.BundleDeploymentOptions) (io.Reader, error) {
+func HelmChart(name string, m *manifest.Manifest, options fleet.BundleDeploymentOptions) (io.Reader, error) {
 	var (
 		style = bundlereader.DetermineStyle(m, options)
 		err   error
@@ -29,10 +38,118 @@ func ToChart(name string, m *manifest.Manifest, options fleet.BundleDeploymentOp
 		}
 	}
 
-	m, err = helm.Process(name, m, style)
+	m, err = process(name, m, style)
 	if err != nil {
 		return nil, err
 	}
 
 	return m.ToTarGZ()
+}
+
+// process filters the manifests resources and adds a Chart.yaml if missing
+func process(name string, m *manifest.Manifest, style bundlereader.Style) (*manifest.Manifest, error) {
+	newManifest := toChart(m, style)
+	if !style.HasChartYAML {
+		return addChartYAML(name, m, newManifest)
+	}
+	return newManifest, nil
+}
+
+func move(m *manifest.Manifest, from, to string) (result []fleet.BundleResource) {
+	if from == "." {
+		from = ""
+	} else if from != "" {
+		from += "/"
+	}
+	for _, resource := range m.Resources {
+		if strings.HasPrefix(resource.Name, from) {
+			resource.Name = to + strings.TrimPrefix(resource.Name, from)
+			result = append(result, resource)
+		}
+	}
+	return result
+}
+
+// manifests returns a filtered list of BundleResources
+// It also treats the 'templates/' directory as a special case.
+func manifests(m *manifest.Manifest) (result []fleet.BundleResource) {
+	var ignorePrefix []string
+	for _, resource := range m.Resources {
+		if fleetyaml.IsFleetYamlSuffix(resource.Name) ||
+			strings.HasSuffix(resource.Name, "/Chart.yaml") {
+			ignorePrefix = append(ignorePrefix, filepath.Dir(resource.Name)+"/")
+		}
+	}
+
+outer:
+	for _, resource := range m.Resources {
+		if fleetyaml.IsFleetYaml(resource.Name) {
+			continue
+		}
+		if !strings.HasSuffix(resource.Name, ".yaml") &&
+			!strings.HasSuffix(resource.Name, ".json") &&
+			!strings.HasSuffix(resource.Name, ".yml") {
+			continue
+		}
+		for _, prefix := range ignorePrefix {
+			if strings.HasPrefix(resource.Name, prefix) {
+				continue outer
+			}
+		}
+		if strings.HasPrefix(resource.Name, "templates/") {
+			resource.Name = "chart/" + resource.Name
+		} else {
+			resource.Name = rawyaml.YAMLPrefix + resource.Name
+		}
+		result = append(result, resource)
+	}
+
+	return result
+}
+
+func toChart(m *manifest.Manifest, style bundlereader.Style) *manifest.Manifest {
+	var (
+		resources []fleet.BundleResource
+	)
+
+	if style.ChartPath != "" {
+		resources = move(m, filepath.Dir(style.ChartPath), "chart/")
+	} else if style.IsRawYAML() {
+		resources = manifests(m)
+	}
+
+	return &manifest.Manifest{
+		Resources: resources,
+		Commit:    m.Commit,
+	}
+}
+
+func addChartYAML(name string, m, newManifest *manifest.Manifest) (*manifest.Manifest, error) {
+	_, hash, err := m.Content()
+	if err != nil {
+		return nil, err
+	}
+
+	if newManifest.Commit != "" && len(newManifest.Commit) > 12 {
+		hash = "git-" + newManifest.Commit[:12]
+	}
+
+	_, chartName := kv.RSplit(name, "/")
+	metadata := chart.Metadata{
+		Name:       chartName,
+		Version:    "v0.0.0+" + hash,
+		APIVersion: "v2",
+	}
+
+	chart, err := yaml.Marshal(metadata)
+	if err != nil {
+		return nil, err
+	}
+
+	newManifest.Resources = append(newManifest.Resources, fleet.BundleResource{
+		Name:    "chart/Chart.yaml",
+		Content: string(chart),
+	})
+
+	return newManifest, nil
 }
